Switch formatDateFromNow() to use moment.humanize()
